Switch back to basic module, and add setupHighlightJs method for config
