+++ app/assets/javascripts/admin/templates/components/highlighted-code.hbs
@@ -1,1 +1,1 @@
-<pre><code class={{lang}}>{{code}}</code></pre>
+<pre><code class={{codeClasses}}>{{displayCode}}</code></pre>
